Home

ConstructionBase.jl

ConstructionBase.constructorof

ConstructionBase.setproperties

ConstructionBase.constructorof — Method.

constructorof(T::Type)

Return an object ctor that can be used to construct objects of type T from their field values. Typically ctor will be the type T with all parameters removed:

julia> struct T{A,B};a::A;b::B;end

julia> constructorof(T{Int,Int})
T

It is however not guaranteed, that ctor is a type at all:

julia> struct S
           a
           b
           checksum
           S(a,b) = new(a,b,a+b)
       end

julia> ConstructionBase.constructorof(S) = (a,b,checksum) -> (@assert a+b == checksum; S(a,b))

julia> constructorof(S)(1,2)
S(1, 2, 3)

Instead ctor can be any object that satisfies the following properties:

It must be possible to reconstruct an object from its fields:

ctor = constructorof(typeof(obj))
@assert obj == ctor(fieldvalues(obj)...)
@assert typeof(obj) == typeof(ctor(fieldvalues(obj)...))

The other direction should hold for as many values of args as possible:

ctor = constructorof(T)
fieldvalues(ctor(args...)) == args

For instance given a suitable parametric type it should be possible to change the type of its fields:

julia> using ConstructionBase: constructorof

julia> struct T{A,B};a::A;b::B;end

julia> t = T(1,2)
T{Int64,Int64}(1, 2)

julia> constructorof(typeof(t))(1.0, 2)
T{Float64,Int64}(1.0, 2)

julia> constructorof(typeof(t))(10, 2)
T{Int64,Int64}(10, 2)

source

ConstructionBase.setproperties — Function.

setproperties(obj, patch)

Return a copy of obj with attributes updates accoring to patch.

Examples

julia> using ConstructionBase

julia> struct S;a;b;c; end

julia> s = S(1,2,3)
S(1, 2, 3)

julia> setproperties(s, (a=10,c=4))
S(10, 2, 4)

julia> setproperties((a=1,c=2,b=3), (a=10,c=4))
(a = 10, c = 4, b = 3)

There is also a convenience method, which builds the patch argument from keywords:

setproperties(obj; kw...)

Examples

julia> struct S;a;b;c; end

julia> o = S(10, 2, 4)
S(10, 2, 4)

julia> setproperties(o, a="A", c="cc")
S("A", 2, "cc")

Overloading

WARNING The signature setproperties(obj::MyType; kw...) should never be overloaded. Instead setproperties(obj::MyType, patch) should be overloaded.

source